requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
xgboost
tensorflow
h2o
autosklearn
